+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/atomastic/strings) (2020-10-28)
+
+* simplify length() method.
+* add lines() method.
+* add words() method.
+* add charsFrequency() method.
+* add wordsFrequency() method.
+* add wordsSortDesc() and wordsSortAsc() methods.
+* add replaceDashes() method.
+* add replacePunctuations() method.
+* add getEncoding() and setEncoding() methods.
+* add replaceNonAlpha() method and update replaceNonAlphanumeric() method.
+* add replaceNonAlphanumeric() method.
+* add isUrl() method.
+* add isEmail() method.
+* improve stripSpaces() method.
+* rewrite method logic and rename countWords() to wordsCount()
+* rename method words() to wordsLimit() and improve tests for this method.
+* update tests for segments() method.
+
+### BREAKING CHANGES
+
+* USE METHOD wordsCount() INSTEAD OF countWords()
+* USE wordsLimit() INSTEAD OF words()
+
 <a name="1.0.0"></a>
 # [1.0.0](https://github.com/atomastic/strings) (2020-09-25)
 * Initial release